Finish Caligor Raid, only missing the Portal back to Unknown Land and NoKill Ending

+++ OpenNos.GameObject/Helpers/UserInterfaceHelper.cs
@@ -37,7 +37,7 @@ namespace OpenNos.GameObject.Helpers
 
         public static string GenerateBSInfo(byte mode, short title, short time, short text) => $"bsinfo {mode} {title} {time} {text}";
 
-        public static string GenerateCHDM(int maxhp, int angeldmg, int demondmg, short time) => $"ch_dm {maxhp} {angeldmg} {demondmg} {time}";
+        public static string GenerateCHDM(int maxhp, int angeldmg, int demondmg, int time) => $"ch_dm {maxhp} {angeldmg} {demondmg} {time}";
 
         public static string GenerateDelay(int delay, int type, string argument) => $"delay {delay} {type} {argument}";
